// client/src/store/ref/locals.ts
// Переменные, хранящиеся в localStorage
export enum loc {
    USER_ID = 'CHALKBOARD_USER_ID', // Id пользователя
    USER_NAME = 'CHALKBOARD_USER_NAME', // Имя пользователя
    USER_COLOR = 'CHALKBOARD_USER_COLOR', // Цвет пользователя
    USER_CANVAS = 'CHALKBOARD_USER_CANVAS', // Холст пользователя
    USER_LAST_BOARD_ID = 'CHALKBOARD_USER_LAST_BOARD_ID', // ID последней доски
    USER_LAST_BOARD_ROLE = 'CHALKBOARD_USER_LAST_BOARD_ROLE', // Роль пользователя на последней доске
}